+++ .gitignore
@@ -101,6 +101,9 @@ ENV/
 .vscode
 settings.json
 
+# PyCharm
+.idea
+
 # mypy
 .mypy_cache/
