+++ search/routes/api/__init__.py
@@ -1,22 +1,12 @@
 """Provides routing blueprint from the search API."""
 
-import json
-from typing import Dict, Callable, Union, Any, Optional, List
-from functools import wraps
-from urllib.parse import urljoin, urlparse, parse_qs, urlencode, urlunparse
+from flask import Blueprint, make_response, request, Response
 
-from flask.json import jsonify
-from flask import Blueprint, make_response, render_template, redirect, \
-    request, Response, url_for
-from werkzeug.urls import Href, url_encode, url_parse, url_unparse, url_encode
-from werkzeug.datastructures import MultiDict, ImmutableMultiDict
-
-from arxiv import status
 from arxiv.base import logging
-from werkzeug.exceptions import InternalServerError
+from search import serialize
 from search.controllers import api
 
-from . import serialize, exceptions, classic
+from search.routes.api import exceptions
 
 from arxiv.users.auth.decorators import scoped
 from arxiv.users.auth import scopes
