+++ search/serialize/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["JSONSerializer", "as_json", "AtomXMLSerializer", "as_atom"]
+
+from search.serialize.json import JSONSerializer, as_json
+from search.serialize.atom import AtomXMLSerializer, as_atom
