+++ search/serialize/base.py
@@ -1,0 +1,2 @@
+class BaseSerializer(object):
+    """Base class for API serializers."""
